move FQName() to Config
