+++ cmd/registry/cmd/diff/diff.go
@@ -20,7 +20,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/apigee/registry/cmd/registry/cmd/util"
 	"github.com/apigee/registry/cmd/registry/core"
 	"github.com/apigee/registry/log"
 	"github.com/apigee/registry/pkg/connection"
@@ -45,7 +44,7 @@ func Command() *cobra.Command {
 				log.FromContext(ctx).WithError(err).Fatal("Failed to get config")
 			}
 			for i := range args {
-				args[i] = util.FQName(c, args[i])
+				args[i] = c.FQName(args[i])
 			}
 
 			client, err := connection.NewRegistryClient(ctx)
