+++ cmd/registry/cmd/export/sheet.go
@@ -21,7 +21,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/apigee/registry/cmd/registry/cmd/util"
 	"github.com/apigee/registry/cmd/registry/core"
 	"github.com/apigee/registry/log"
 	"github.com/apigee/registry/pkg/connection"
@@ -50,7 +49,7 @@ func sheetCommand() *cobra.Command {
 				log.FromContext(ctx).WithError(err).Fatal("Failed to get config")
 			}
 			for i := range args {
-				args[i] = util.FQName(c, args[i])
+				args[i] = c.FQName(args[i])
 			}
 
 			var path string
