+++ cmd/registry/cmd/label/label.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/apigee/registry/cmd/registry/cmd/util"
 	"github.com/apigee/registry/cmd/registry/core"
 	"github.com/apigee/registry/gapic"
 	"github.com/apigee/registry/log"
@@ -46,7 +45,7 @@ func Command() *cobra.Command {
 			if err != nil {
 				log.FromContext(ctx).WithError(err).Fatal("Failed to get config")
 			}
-			args[0] = util.FQName(c, args[0])
+			args[0] = c.FQName(args[0])
 
 			client, err := connection.NewRegistryClient(ctx)
 			if err != nil {
